refactor(relay): extract shared rate limiter, attachment helpers, and response headers
Deduplicate identical implementations across 5 relay proxy modules:

- shared-rate-limiter.ts: SlidingWindowRateLimiter class replacing inline
  Map+checkRateLimit+setInterval in anthropic-proxy, git-proxy, http-credential-proxy
- attachment-helpers.ts: MIME_EXTENSION_MAP, sanitizeFilename, buildAttachmentFilename,
  ensureDocumentsDir replacing duplicates in provider-proxy and capabilities
- proxy-headers.ts: EXCLUDED_RESPONSE_HEADERS set and forwardResponseHeaders()
  replacing inline header filtering in anthropic-proxy, git-proxy, http-credential-proxy

Removes ~230 lines of duplicated code. No behavioral changes.
