+++ src/relay/capabilities.ts
@@ -1,4 +1,3 @@
-import crypto from "node:crypto";
 import fs from "node:fs";
 import http from "node:http";
 import path from "node:path";
@@ -37,6 +36,7 @@ import type { RuntimeSnapshot } from "../system-metadata.js";
 import { buildRuntimeSnapshot } from "../system-metadata.js";
 import { sendAdminAlert } from "../telegram/admin-alert.js";
 import { handleAnthropicProxyRequest, isAnthropicProxyRequest } from "./anthropic-proxy.js";
+import { buildAttachmentFilename, ensureDocumentsDir } from "./attachment-helpers.js";
 import { type ProviderProxyRequest, proxyProviderRequest } from "./provider-proxy.js";
 import {
 	handleTokenExchange,
@@ -344,44 +344,6 @@ function parseUserId(input: unknown): { userId?: string; error?: string } {
 		return { error: "userId too long." };
 	}
 	return { userId: trimmed };
-}
-
-const MIME_EXTENSION_MAP: Record<string, string> = {
-	"application/pdf": ".pdf",
-	"image/png": ".png",
-	"image/jpeg": ".jpg",
-	"image/jpg": ".jpg",
-	"image/webp": ".webp",
-};
-
-function sanitizeFilename(input?: string): string {
-	if (!input || typeof input !== "string") {
-		return "attachment";
-	}
-	const base = path.basename(input.trim());
-	const normalized = base.replace(/[^a-zA-Z0-9._-]+/g, "_").replace(/^_+|_+$/g, "");
-	if (!normalized || normalized === "." || normalized === "..") {
-		return "attachment";
-	}
-	return normalized;
-}
-
-function buildAttachmentFilename(filename?: string, mimeType?: string): string {
-	const safe = sanitizeFilename(filename);
-	const extFromName = path.extname(safe);
-	const fallbackExt = mimeType ? (MIME_EXTENSION_MAP[mimeType] ?? "") : "";
-	const ext = extFromName || fallbackExt;
-	const stem = (extFromName ? safe.slice(0, -extFromName.length) : safe) || "attachment";
-	const suffix = `${Date.now()}-${crypto.randomBytes(4).toString("hex")}`;
-	const truncatedStem = stem.slice(0, 80);
-	return `${truncatedStem}-${suffix}${ext}`;
-}
-
-async function ensureDocumentsDir(): Promise<string> {
-	const outboxRoot = getMediaOutboxDirSync();
-	const documentsDir = path.join(outboxRoot, "documents");
-	await fs.promises.mkdir(documentsDir, { recursive: true, mode: 0o700 });
-	return documentsDir;
 }
 
 /**
